move generator to testing directory, add doc
